fixed a couple paths related to renaming the serverdb/db file to database.js and added user stories link to figma in MVP
