+++ MVP.md
@@ -2,10 +2,20 @@
 
 - A Mental Wellness App that starts out using a form that prompts a user to select how they're feeling from a set of options and places them in a room. Our goal is to have a minimum of 2 designs for rooms where users can chat with other users through their avatars and another room with a single-player game. We'll be using Socket.io, Phaser and our stack.
 
-### Technical Challenges:
+## Technical Challenges:
 
 - At least 2 users subscribed to same socket.
 - implementing the chat room feature.
+
+## POC:
+
+- 1 multiplayer room and 1 single player room (single player room could be room with journal)
+- A working Phaser game
+- responding to sockets.
+
+### User Stories:
+
+[figma](https://www.figma.com/file/K8xo5xO9U0VnnQ1co5SHF6/Capstone-Cerulean-Pages?node-id=2%3A892)
 
 ### Stretch Goals:
